Cria opção de partida com identificador aleatório

+++ src/main/java/br/ufjf/dcc193/atividade9/login/PlayMoney.java
@@ -2,10 +2,6 @@ package br.ufjf.dcc193.atividade9.login;
 
 import java.util.Date;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.Pattern;
-
 import org.springframework.format.annotation.DateTimeFormat;
 
 import lombok.Data;
@@ -26,5 +22,5 @@ public class PlayMoney {
     @DateTimeFormat(pattern = "HH:mm:ss")
     private Date hora = new Date();
 
-    private Float montante;
+    private Double montante;
 }
